requirements.txt
openai
yfinance
python-dotenv
faiss
requests
tinydb